Move file to src, to only use exported things
