+++ lib/rounded_picker.dart
@@ -1,6 +1,7 @@
 library flutter_rounded_date_picker;
 
-export './flutter_rounded_date_picker_widget.dart' show showRoundedDatePicker;
-export './flutter_rounded_time_picker_widget.dart' show showRoundedTimePicker;
-export './cupertino_rounded_date_picker.dart' show CupertinoRoundedDatePicker;
-export './cupertino_rounded_duration_picker.dart' show CupertinoRoundedDurationPicker;
+export './src/flutter_rounded_date_picker_widget.dart' show showRoundedDatePicker;
+export './src/flutter_rounded_time_picker_widget.dart' show showRoundedTimePicker;
+export './src/cupertino_rounded_date_picker.dart' show CupertinoRoundedDatePicker;
+export './src/cupertino_rounded_duration_picker.dart' show CupertinoRoundedDurationPicker;
+export './src/era_mode.dart' show EraMode;
