+++ lib/src/cupertino_rounded_date_picker.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_rounded_date_picker/era_mode.dart';
-import 'package:flutter_rounded_date_picker/flutter_cupertino_rounded_date_picker_widget.dart';
+import 'package:flutter_rounded_date_picker/src/era_mode.dart';
+import 'package:flutter_rounded_date_picker/src/flutter_cupertino_rounded_date_picker_widget.dart';
 
 class CupertinoRoundedDatePicker {
   static show(
